refactor: changes queue mechanism to use worker

+++ app/src/main/java/com/kelsos/mbrc/common/data/Repository.kt
@@ -10,5 +10,5 @@ interface Repository<T : Any> {
   fun search(term: String): Flow<PagingData<T>>
   suspend fun cacheIsEmpty(): Boolean
   suspend fun count(): Long
-  suspend fun getById(id: Int): T?
+  suspend fun getById(id: Long): T?
 }
